Moved the oauth token service to the tools folder
