--- src/SubscribePro/Service/OauthToken/OauthToken.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace SubscribePro\Service\OauthToken;
-
-use SubscribePro\Service\DataObject;
-
-class OauthToken extends DataObject implements OauthTokenInterface {}
-
