--- src/SubscribePro/Service/OauthToken/OauthTokenInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace SubscribePro\Service\OauthToken;
-
-use SubscribePro\Service\DataInterface;
-
-interface OauthTokenInterface extends DataInterface {}
-
